dependency/fmdm-simulator2/app/requirements.txt
# Pillow==7.1.1
numpy<2.0
Pillow>=9.1,<11
scikit-image==0.23.2
pluggy==0.13.1
py==1.8.1
pytest==5.4.1
transforms3d==0.3.1
gym==0.15.4
pybullet
# scikit-image
pytransform3d
ftfy
regex
tqdm